+++ lib/shared/widgets/triangle-clipper.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/cupertino.dart';
+
+// Renders an isosceles triangle by clipping a Container.
+// One usage is in tooltips for arrow indicator.
+class TriangleClipper extends CustomClipper<Path> {
+  @override
+  Path getClip(Size size) {
+    final path = Path();
+    path.moveTo(size.width / 2, 0);
+    path.lineTo(0, size.height);
+    path.lineTo(size.width, size.height);
+    path.close();
+
+    return path;
+  }
+
+  @override
+  bool shouldReclip(TriangleClipper oldClipper) => false;
+}
+